+++ src/pages/redirectflow/Confirm.vue
@@ -10,10 +10,10 @@ import { useEstimateChanges } from "@/components/payments/EstimateChangesComposa
 import PermissionsTx from "@/components/permissionsTx/PermissionsTx.vue";
 import { TransactionChannelDataType } from "@/utils/enums";
 import { hideCrispButton, openCrispChat } from "@/utils/helpers";
-import { DecodedDataType, decodeInstruction } from "@/utils/instructionDecoder";
+import { decodeAllInstruction, DecodedDataType, decodeInstruction } from "@/utils/instructionDecoder";
 import { FinalTxData } from "@/utils/interfaces";
 import { redirectToResult, useRedirectFlow } from "@/utils/redirectflowHelpers";
-import { calculateTxFee, decodeAllInstruction, parsingTransferAmount } from "@/utils/solanaHelpers";
+import { calculateTxFee, parsingTransferAmount } from "@/utils/solanaHelpers";
 
 import ControllerModule from "../../modules/controllers";
